+++ _scripts/example.ts
@@ -1,7 +1,11 @@
-// deno run --allow-write example.ts
+// deno run --allow-write _scripts/example.ts
 import { Changelog, Release } from "../mod.ts";
 
 const changelog = new Changelog("Changelog")
+  .addRelease(
+    new Release("2.6.2", "2025-03-23")
+      .added("New option `--latest-release-full` to print the full release #59")
+  )
   .addRelease(
     new Release("2.6.1", "2025-02-24")
       .fixed("NPM publishing #55, #56.")