+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [2.6.2] - 2025-03-23
 ### Added
-- New option `--latest-release-full` to print the full release
+- New option `--latest-release-full` to print the full release [#59]
 
 ## [2.6.1] - 2025-02-24
 ### Fixed
@@ -123,6 +123,7 @@ New version merging Deno and Node code using Deno's `dnt` package.
 [#51]: https://github.com/oscarotero/keep-a-changelog/issues/51
 [#55]: https://github.com/oscarotero/keep-a-changelog/issues/55
 [#56]: https://github.com/oscarotero/keep-a-changelog/issues/56
+[#59]: https://github.com/oscarotero/keep-a-changelog/issues/59
 
 [2.6.2]: https://github.com/oscarotero/keep-a-changelog/compare/v2.6.1...v2.6.2
 [2.6.1]: https://github.com/oscarotero/keep-a-changelog/compare/v2.6.0...v2.6.1